pyproject.toml
[project]
name = "process-gpt-execution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "a2a-sdk==0.2.4",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.13",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anthropic==0.54.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "blinker==1.9.0",
    "build==1.2.2.post1",
    "cachecontrol==0.14.3",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.6.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "contextvars==2.4",
    "cryptography==45.0.4",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "deprecation==2.1.0",
    "distlib==0.3.9",
    "distro==1.9.0",
    "docker==7.1.0",
    "durationpy==0.10",
    "fastapi==0.115.12",
    "filelock==3.18.0",
    "filetype==1.2.0",
    # "firebase-admin==6.9.0",  # FCM 서비스로 분리됨
    "requests==2.31.0",  # FCM 클라이언트에서 사용
    "flask==3.1.1",
    "flatbuffers==25.2.10",
    "flupy==1.2.2",
    "frozenlist==1.7.0",
    "fsspec==2025.5.1",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.1",
    "google-api-python-client==2.172.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-cloud-core==2.4.3",
    "google-cloud-firestore==2.21.0",
    "google-cloud-storage==3.1.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "gotrue==2.12.0",
    "greenlet==3.2.3",
    "grpcio==1.73.0",
    "grpcio-reflection==1.71.0",
    "grpcio-status==1.71.0",
    "grpcio-tools==1.71.0",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.33.0",
    "humanfriendly==10.0",
    "hyperframe==6.1.0",
    "idna==3.10",
    "immutables==0.21",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "iniconfig==2.1.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kubernetes==33.1.0",
    "langchain==0.3.25",
    "langchain-anthropic==0.3.15",
    "langchain-community==0.3.25",
    "langchain-core==0.3.65",
    "langchain-google-genai==2.1.5",
    "langchain-ollama==0.3.3",
    "langchain-openai==0.3.23",
    "langchain-text-splitters==0.3.8",
    "langserve==0.3.1",
    "langsmith==0.3.45",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mem0ai==0.1.108",
    "mmh3==5.1.0",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "msgpack==1.1.1",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "numpy==2.3.0",
    "oauthlib==3.2.2",
    "ollama==0.5.1",
    "onnxruntime==1.22.0",
    "openai==1.86.0",
    "opentelemetry-api==1.34.1",
    "opentelemetry-exporter-otlp-proto-common==1.34.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.34.1",
    "opentelemetry-instrumentation==0.55b1",
    "opentelemetry-instrumentation-asgi==0.55b1",
    "opentelemetry-instrumentation-fastapi==0.55b1",
    "opentelemetry-proto==1.34.1",
    "opentelemetry-sdk==1.34.1",
    "opentelemetry-semantic-conventions==0.55b1",
    "opentelemetry-util-http==0.55b1",
    "orjson==3.10.18",
    "overrides==7.7.0",
    "packaging==24.2",
    "pgvector==0.3.6",
    "pipenv==2025.0.3",
    "platformdirs==4.3.8",
    "pluggy==1.6.0",
    "portalocker==2.10.1",
    "postgrest==1.0.2",
    "posthog==4.10.0",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pyparsing==3.2.3",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyreadline3==3.5.4",
    "pytest==8.4.0",
    "pytest-mock==3.14.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "qdrant-client==1.14.2",
    "realtime==2.4.3",
    "regex==2024.11.6",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sse-starlette==2.3.6",
    "starlette==0.46.2",
    "storage3==0.11.3",
    "strenum==0.4.15",
    "supabase==2.15.3",
    "supafunc==0.9.4",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "transformers==4.52.4",
    "typer==0.16.0",
    "typing-extensions==4.14.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "uritemplate==4.2.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.3",
    "vecs==0.4.5",
    "virtualenv==20.31.2",
    "watchfiles==1.0.5",
    "websocket-client==1.8.0",
    "websockets==14.2",
    "werkzeug==3.1.3",
    "wrapt==1.17.2",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.23.0",
]
